requirements.txt
paddlepaddle
paddleocr>=2.9.1
opencv-python
numpy
Pillow
imagehash
pandas
tqdm
scikit-image
matplotlib